Updates auth method, makes test and makes psswrd utils
