+++ routes/teachers.js
@@ -1,37 +1,12 @@
-var express = require('express');
-var router = express.Router();
-var crypto = require('crypto');
-var mongoose = require('mongoose');
-var Teachers = require('../models/teacher.model');
+const express = require('express');
+const router = express.Router();
+const mongoose = require('mongoose');
 
-var authMiddleware = require('../auth.middleware');
+const Teachers = require('../models/teacher.model');
+const hashPassword = require('../utils/password');
+const authMiddleware = require('../auth.middleware');
 
 mongoose.connect('mongodb://db:27017/base');
-
-/* generates salt for hash with random char string */
-const genRandomString = function genRandomString(length) {
-  return crypto.randomBytes(Math.ceil(length / 2))
-    .toString('hex')
-    .slice(0, length);
-};
-
-/* hash password with sha512 */
-const sha512 = function sha512(password, salt) {
-  const hash = crypto.createHmac('sha512', salt);
-  hash.update(password);
-  const value = hash.digest('hex');
-  return {
-    salt,
-    passwordHash: value,
-  };
-};
-
-function hashPassword(password) {
-  const salt = genRandomString(16);
-  const passwordData = sha512(password, salt);
-
-  return passwordData;
-}
 
 /* POST teachers. */
 router.post('/new', (req, res, next) => {
